refactor: change try/catch error handling

+++ client/src/actions/comments/delete-comment.ts
@@ -4,13 +4,14 @@ import { ApiResponse } from "@/interfaces/actions"
 import ApiError from "@/utils/api-error"
 import { PHOTO_DELETE_COMMENT } from "@/utils/api-urls"
 import { revalidateTag } from "next/cache"
-import verifyAuthenticationAndReturnToken from "@/utils/verify-authentication"
+import { cookies } from "next/headers"
 
 export default async function deleteComment(photoId: string, commentId: string) {
     const URL = PHOTO_DELETE_COMMENT(photoId, commentId)
+    const auth_token = cookies().get("auth_token")?.value
 
     try {
-        const auth_token = await verifyAuthenticationAndReturnToken()
+        if (!auth_token) return ApiError(new Error("You are not authenticated."))
 
         const response = await fetch(URL, {
             method: "DELETE",
@@ -20,7 +21,7 @@ export default async function deleteComment(photoId: string, commentId: string) 
 
         const responseData = (await response.json()) as ApiResponse
 
-        if (!response.ok) throw new Error(responseData.message ?? "An unknown error occurred.")
+        if (!response.ok) return ApiError(new Error(responseData.message ?? "An unknown error occurred."))
 
         revalidateTag("comment")
         return {
