+++ client/src/actions/photos/delete-photo.ts
@@ -3,13 +3,14 @@
 import { ApiResponse } from "@/interfaces/actions"
 import ApiError from "@/utils/api-error"
 import { PHOTO_DELETE_PHOTO } from "@/utils/api-urls"
-import verifyAuthenticationAndReturnToken from "@/utils/verify-authentication"
+import { cookies } from "next/headers"
 
 export default async function deletePhoto(photoId: string) {
     const URL = PHOTO_DELETE_PHOTO(photoId)
 
+    const auth_token = cookies().get("auth_token")?.value
     try {
-        const auth_token = await verifyAuthenticationAndReturnToken()
+        if (!auth_token) return ApiError(new Error("You are not authenticated."))
 
         const response = await fetch(URL, {
             method: "DELETE",
@@ -19,7 +20,7 @@ export default async function deletePhoto(photoId: string) {
 
         const responseData = (await response.json()) as ApiResponse
 
-        if (!response.ok) throw new Error(responseData.message ?? "An unknown error occurred.")
+        if (!response.ok) return ApiError(new Error(responseData.message ?? "An unknown error occurred."))
 
         return {
             ok: true,
